safari-600.7-branch/LayoutTests/inspector-obsolete/http-tests/compiler-source-mapping-debug-expected.txt

CONSOLE MESSAGE: line 2: button clicked!
Tests installing compiler source map in scripts panel.


Debugger was enabled.

Running: testSetBreakpoint
Page reloaded.
Script source was shown.
Script execution paused.
Call stack:
    0) handleClick (compiled.js:2)
    1) clickButton (compiler-source-mapping-debug.html:11)
Script execution resumed.
Debugger was disabled.
